backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
google-genai
google-adk
python-dotenv==1.0.1
dotenv
fastapi
uvicorn
websockets
